#425 show version in app #424 fix androidx conversion bug (create activity progress was int instead of double #430 Reduce fotn text size

+++ app/src/main/java/org/sil/storyproducer/controller/export/CreateActivity.kt
@@ -56,7 +56,7 @@ class CreateActivity : PhaseBaseActivity() {
                 return@Runnable
             }
 
-            var progress: Int
+            var progress: Double
             synchronized(storyMakerLock) {
                 //Stop if storyMaker was cancelled by someone else.
                 if (storyMaker == null) {
@@ -64,7 +64,7 @@ class CreateActivity : PhaseBaseActivity() {
                     return@Runnable
                 }
 
-                progress = storyMaker!!.progress.toInt()
+                progress = storyMaker!!.progress
                 isDone = storyMaker!!.isDone
             }
             updateProgress((progress * PROGRESS_MAX).toInt())
